+++ src/lang/pl/builder.php
@@ -10,14 +10,15 @@
  */
 return [
 
-	'ok'                   => 'OK',
-	'no_error_message_fmt' => 'Błąd #:error_code',
+	'ok'                           => 'OK',
+	'no_error_message'             => 'Błąd #:error_code',
 
 	// can be used by Exception Handler (if enabled)
-	'uncaught_exception_fmt'  => 'Nieprzechwycony wyjątek: :message',
-	'unknown_method'          => 'Nieznana metoda',
-	'http_exception_fmt'      => 'Wyjątek HTTP: :message',
-	'service_in_maintenance'  => 'Trwa przerwa serwisowa',
+	'uncaught_exception'           => 'Nieprzechwycony wyjątek: :message',
+	'uncaught_exception_no_prefix' => ':message',
+	'http_not_found'               => 'Nieznana metoda',
+	'http_exception'               => 'Wyjątek HTTP: :message',
+	'http_service_unavailable'     => 'Trwa przerwa serwisowa',
 
 ];