+++ CHANGES.md
@@ -2,10 +2,13 @@
 
 ## CHANGE LOG ##
 
-* v1.5.1 (2016-04-21)
+* v2.0.0 (2016-04-21)
+   * Configuration keys changed
+   * Built-in localization keys changed
    * Added errorWithMessageAndData()
    * Added env('EX_INCLUDE_CLASS_NAME') to control include_class_name exception handler feature
-   * If app runs in DEBUG mode, ExceptionHandlerHelper now adds 'file' and 'line' to returned JSON
+   * If app runs in DEBUG mode, ExceptionHandlerHelper now adds 'file' and 'line' to returned error JSONs
+   * ExceptionHandlerHelper now automatically resolves message mappings and needs no config entries
 
 * v1.5.0 (2016-04-18)
    * ExHandler: ExceptionHandler is now replaced by ExceptionHandlerHelper